+++ Front-end/src/components/users/Modify.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const Div = styled.div`
+  width: auto;
+  margin-left: 400px;
+`;
+const Modify = () => {
+  return <Div>modify</Div>;
+};
+
+export default Modify;
