+++ Front-end/src/components/users/MyMinute.jsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+import React from 'react';
+
+const Div = styled.div`
+  width: auto;
+  margin-left: 400px;
+`;
+
+const MyMinute = () => {
+  return (
+    <>
+      <Div>my Minute</Div>
+    </>
+  );
+};
+
+export default MyMinute;
