+++ Front-end/src/pages/Minutes/index.jsx
@@ -28,9 +28,9 @@ const Minutes = () => {
     <>
       <Header />
       <ContainerWrap>
-        <h2 style={{fontSize:'25px',fontWeight:'bold',padding:'20px 0'}}>회의록 작성</h2>
+        <h2 style={{ fontSize: '25px', fontWeight: 'bold', padding: '20px 0' }}>회의록 작성</h2>
         <Container>
-          <form onSubmit={onSubmit} style={{width:'100%'}}>
+          <form onSubmit={onSubmit} style={{ width: '100%' }}>
             <SpanDiv>
               <Span>회의 일자</Span>
               <InputLeft type="date" className="date"></InputLeft>
@@ -45,7 +45,7 @@ const Minutes = () => {
             </SpanDiv>
             <SpanDiv>
               <ContentDiv>내용</ContentDiv>
-              <div style={{width:'100%',textAlign:'center'}}>
+              <div style={{ width: '100%', textAlign: 'center' }}>
                 <InputText type="text" className="content"></InputText>
               </div>
             </SpanDiv>
@@ -59,7 +59,7 @@ const Minutes = () => {
               <InputCheck type="checkbox" name="name2"></InputCheck>
               <Label for="name2">비공개</Label>
               <InputCheck type="checkbox" name="name1"></InputCheck>
-              <Select name="team">
+              <Select name="team" style={{ fontSize: '13px' }}>
                 <option value="경영지원팀">경영지원팀</option>
                 <option value="영업팀">영업팀</option>
                 <option value="인사팀">인사팀</option>
@@ -67,11 +67,19 @@ const Minutes = () => {
             </SpanDiv>
           </form>
         </Container>
-        <div style={{width:'80%',display:'flex',justifyContent:'space-between',marginTop:'20px',maxWidth:'800px'}}>
-          <Link to="/minutes" style={{width:'45%'}}>
+        <div
+          style={{
+            width: '80%',
+            display: 'flex',
+            justifyContent: 'space-between',
+            marginTop: '20px',
+            maxWidth: '800px',
+          }}
+        >
+          <Link to="/minutes" style={{ width: '45%' }}>
             <ButtonConfirm type="submit">작성하기</ButtonConfirm>
           </Link>
-          <Link to="/minutes"  style={{width:'45%'}}>
+          <Link to="/minutes" style={{ width: '45%' }}>
             <ButtonCancel type="submit">취소</ButtonCancel>
           </Link>
         </div>
